+++ lib/crowd_control/provider/docker.ex
@@ -82,6 +82,7 @@ defmodule CrowdControl.Provider.Docker do
   alias CrowdControl.Backend.Sandboxd.API, as: AgentAPI
   alias CrowdControl.Provider
   alias CrowdControl.Provider.Endpoint
+  alias CrowdControl.ReqAdapter
   alias CrowdControl.Store
 
   @default_agent_port 8080
@@ -299,15 +300,8 @@ defmodule CrowdControl.Provider.Docker do
     %Endpoint{
       base_url: "http://127.0.0.1:#{host_port}",
       token: Provider.token(handle.session_key),
-      req_options: req_options(handle)
+      req_options: ReqAdapter.req_options(handle.config[:req_adapter])
     }
-  end
-
-  defp req_options(handle) do
-    case handle.config[:req_adapter] do
-      nil -> []
-      adapter -> [adapter: adapter]
-    end
   end
 
   @doc """